+++ pyAMARES/__init__.py
@@ -1,8 +1,8 @@
 __author__ = "Jia Xu, MR Research Facility, University of Iowa"
-__version__ = "0.3.9-dev"
+__version__ = "0.4.0-dev"
 
-print("Current pyAMARES version is %s" % __version__)
-print("Author: %s" % __author__)
+# print("Current pyAMARES version is %s" % __version__)
+# print("Author: %s" % __author__)
 
 from .kernel import *
 from .util import *