+++ CHANGELOG.rst
@@ -1,9 +1,23 @@
 This document describes all notable changes to pyAMARES.
 
+v0.4.0
+------
+**Added**
+- Added the ``initialize_with_lm`` option to both ``fitAMARES`` and ``run_parallel_fitting_with_progress`` functions.
+- Added a ``highlight_dataframe`` function that highlights rows in a DataFrame based on the values of a specified column.
+
+
+
+**Fixed**
+- Updated docstrings in numerous functions to ensure they render properly.
+
 v0.3.9
 ------
 **Added**
 - The peak-wise Signal-to-Noise Ratio (SNR) is now added to each ``result_pd``. The Standard Deviation (SD) of the noise is obtained from the last 10% of points in the FID.
+
+**Fixed**
+- Mute `__version__` and `__author__` printouts. 
 
 v0.3.8 
 ------